# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WetechiNetMon Incident API",
    "description": "Incident management REST API. Errors are RFC 9457 problem details with a stable `error` code.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.0.0"
  },
  "paths": {
    "/api/v1/incidents": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "The caller's tenant's incidents, newest first by default.",
        "operationId": "list_incidents",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "description": "Repeatable. open, acknowledged, investigating, monitoring, recovering, resolved, closed.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "severity",
            "in": "query",
            "description": "Repeatable. info, minor, major, critical.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "priority",
            "in": "query",
            "description": "Repeatable. P1, P2, P3, P4.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "direction",
            "in": "query",
            "description": "incoming, outgoing, internal, other, unknown.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target_type",
            "in": "query",
            "description": "host, network, hostgroup.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "opened_from",
            "in": "query",
            "description": "RFC 3339 UTC, inclusive. Needs `opened_to`; at most 90 days apart.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "opened_to",
            "in": "query",
            "description": "RFC 3339 UTC, exclusive.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "opened_at (default) or last_detected_at.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "desc (default) or asc.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "1 to 200; default 50. Larger values are capped at 200.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "`next_cursor` from the previous page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include_total",
            "in": "query",
            "description": "`true` to count every match; charged twice against the read limit.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of incidents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentPage"
                }
              }
            }
          },
          "400": {
            "description": "An unknown parameter or value, a bad range or cursor",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "401": {
            "description": "No usable token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "The role lacks incident.list",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Opens an incident by hand. Needs `incident.create`.",
        "operationId": "create_incident",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "16 to 255 characters; a retry with the same key and body answers with the incident the first call opened",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIncidentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The incident opened, or on a replay the one the first call opened",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "The incident's URL"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentView"
                }
              }
            }
          },
          "400": {
            "description": "Bad header or body, or an unknown field",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "401": {
            "description": "No usable token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "The role lacks incident.create",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "An incident for the target is already active (its incident_id is given), the key was reused, or a limit is reached",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "413": {
            "description": "The body is over 64 KiB",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "415": {
            "description": "The body is not application/json",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "The title, description or target is invalid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "One incident.",
        "operationId": "get_incident",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "description": "The incident's UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The incident",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentView"
                }
              }
            }
          },
          "400": {
            "description": "The id is not a UUID",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "401": {
            "description": "No usable token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "The role lacks incident.read",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such incident in the caller's tenant",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/incidents/{incident_id}/detections": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "The detection events linked to an incident, oldest first.",
        "operationId": "detections",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "description": "The incident's UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "1 to 500; default 100. Larger values are capped at 500.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "`next_cursor` from the previous page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of linked detection events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionPage"
                }
              }
            }
          },
          "400": {
            "description": "Bad id, parameter or cursor",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "401": {
            "description": "No usable token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "The role lacks incident.read",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such incident in the caller's tenant",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/incidents/{incident_id}/priority": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Changes priority. Needs `incident.priority.change`.",
        "operationId": "priority",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "description": "The incident's UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "16 to 255 characters; a retry with the same key and body replays the first outcome",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriorityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The incident after the change, or the current incident on a replay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentView"
                }
              }
            }
          },
          "400": {
            "description": "Bad id, header or body, or an unknown field",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "401": {
            "description": "No usable token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "The role lacks the permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such incident in the caller's tenant",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict, illegal transition, unchanged state or key reuse",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "413": {
            "description": "The body is over 64 KiB",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "415": {
            "description": "The body is not application/json",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Well-formed but semantically invalid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateIncidentRequest": {
        "type": "object",
        "required": [
          "title",
          "severity",
          "target_scope",
          "target",
          "direction"
        ],
        "properties": {
          "address_family": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "4 or 6. Taken from `target` when it is an address or network;\nrequired for `hostgroup_total`.",
            "minimum": 0
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Up to 8000 characters."
          },
          "direction": {
            "$ref": "#/components/schemas/DirectionValue"
          },
          "priority": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PriorityValue",
                "description": "Defaults from the severity: critical P1, major P2, minor P3, info P4."
              },
              {
                "type": "null"
              }
            ]
          },
          "severity": {
            "$ref": "#/components/schemas/SeverityValue"
          },
          "target": {
            "type": "string",
            "description": "An address, a network in CIDR form, or a hostgroup name."
          },
          "target_scope": {
            "$ref": "#/components/schemas/TargetScope"
          },
          "title": {
            "type": "string",
            "description": "Up to 200 characters."
          }
        },
        "additionalProperties": false
      },
...
      "IncidentPage": {
        "type": "object",
        "description": "A page of incidents.",
        "required": [
          "items",
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IncidentSummaryView"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pass as `cursor` for the next page; `null` on the last."
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Every match, only with `include_total=true`."
          }
        }
      },
...
      "PriorityRequest": {
        "type": "object",
        "required": [
          "expected_version",
          "priority"
        ],
        "properties": {
          "expected_version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "priority": {
            "$ref": "#/components/schemas/PriorityValue"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDocument": {
        "type": "object",
        "description": "RFC 9457 problem details, as every error response carries them.",
        "required": [
          "type",
          "title",
          "status",
          "error",
          "request_id"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "description": "The stable machine-readable code; see `docs/api/error-codes.md`."
          },
          "request_id": {
            "type": "string",
            "description": "The UUIDv7 also returned in `X-Request-Id`."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "`https://wetechi.com/probs/` and the code, as an identifier."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API token, `wnm_` and 64 hex characters, from `wetechinetmon-api token create`."
      }
    }
  },
...
}